// react-app/src/hooks/CartContext.d.ts
import type { ReactNode } from 'react';
export interface CartItem {
    id: string;
    name: string;
    price: number;
    image: string;
    quantity: number;
}
export interface CartContextType {
    cart: CartItem[];
    addToCart: (item: Omit<CartItem, 'quantity'>) => void;
    removeFromCart: (id: string) => void;
    increaseQuantity: (id: string) => void;
    decreaseQuantity: (id: string) => void;
    clearCart: () => void;
    totalItems: number;
}
export interface CartProviderProps {
    children: ReactNode;
}
export declare const CartContext: import("react").Context<CartContextType | undefined>;